+++ src/http_frontend/entry_api/read_single.rs
@@ -3,22 +3,22 @@
 //    License, v. 2.0. If a copy of the MPL was not distributed with this
 //    file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-use crate::auth::policy::{ReadAccessPolicy, WriteAccessPolicy};
+use crate::auth::policy::ReadAccessPolicy;
 use crate::core::status::HttpError;
 use crate::http_frontend::entry_api::{check_and_extract_ts_or_query_id, MethodExtractor};
 use crate::http_frontend::middleware::check_permissions;
 use crate::http_frontend::HttpServerState;
 use crate::storage::bucket::Bucket;
-use crate::storage::entry::Labels;
+
 use crate::storage::reader::RecordReader;
-use crate::storage::writer::{Chunk, RecordWriter};
+
 use axum::body::StreamBody;
-use axum::extract::{BodyStream, Path, Query, State};
-use axum::headers::{Expect, Header, HeaderMap, HeaderName};
+use axum::extract::{Path, Query, State};
+use axum::headers::{Header, HeaderMap, HeaderName};
 use axum::response::IntoResponse;
 use bytes::Bytes;
 use futures_util::{Stream, StreamExt};
-use log::{debug, error};
+
 use std::collections::HashMap;
 use std::pin::Pin;
 use std::sync::{Arc, RwLock};
@@ -151,7 +151,7 @@ mod tests {
     use axum::extract::FromRequest;
     use axum::http::Request;
 
-    use crate::http_frontend::entry_api::tests::{components, path};
+    use crate::http_frontend::tests::{components, headers, path_to_entry_1};
     use crate::storage::query::base::QueryOptions;
     use rstest::*;
     use std::path::PathBuf;
@@ -163,18 +163,19 @@ mod tests {
     #[tokio::test]
     async fn test_single_read_ts(
         components: Arc<RwLock<HttpServerState>>,
-        path: Path<HashMap<String, String>>,
+        path_to_entry_1: Path<HashMap<String, String>>,
+        headers: HeaderMap,
         #[case] method: String,
         #[case] body: String,
     ) {
         let mut response = read_single_record(
             State(Arc::clone(&components)),
-            path,
+            path_to_entry_1,
             Query(HashMap::from_iter(vec![(
                 "ts".to_string(),
                 "0".to_string(),
             )])),
-            HeaderMap::new(),
+            headers,
             MethodExtractor::new(&method),
         )
         .await
@@ -198,20 +199,32 @@ mod tests {
     #[tokio::test]
     async fn test_single_read_query(
         components: Arc<RwLock<HttpServerState>>,
-        path: Path<HashMap<String, String>>,
+        path_to_entry_1: Path<HashMap<String, String>>,
+        headers: HeaderMap,
         #[case] method: String,
         #[case] body: String,
     ) {
-        let query_id = query_records(&components);
+        let query_id = {
+            components
+                .write()
+                .unwrap()
+                .storage
+                .get_bucket(path_to_entry_1.get("bucket_name").unwrap())
+                .unwrap()
+                .get_entry(path_to_entry_1.get("entry_name").unwrap())
+                .unwrap()
+                .query(0, u64::MAX, QueryOptions::default())
+                .unwrap()
+        };
 
         let mut response = read_single_record(
             State(Arc::clone(&components)),
-            path,
+            path_to_entry_1,
             Query(HashMap::from_iter(vec![(
                 "q".to_string(),
                 query_id.to_string(),
             )])),
-            HeaderMap::new(),
+            headers,
             MethodExtractor::new(&method),
         )
         .await
@@ -227,28 +240,5 @@ mod tests {
             response.data().await.unwrap_or(Ok(Bytes::new())).unwrap(),
             Bytes::from(body)
         );
-    }
-
-    fn query_records(components: &Arc<RwLock<HttpServerState>>) -> u64 {
-        let query_id = components
-            .write()
-            .unwrap()
-            .storage
-            .get_bucket("bucket-1")
-            .unwrap()
-            .get_entry("entry-1")
-            .unwrap()
-            .query(
-                0,
-                1,
-                QueryOptions {
-                    continuous: false,
-                    include: HashMap::new(),
-                    exclude: HashMap::new(),
-                    ttl: Duration::from_secs(1),
-                },
-            )
-            .unwrap();
-        query_id
     }
 }
